Added unit tests for "df_chat" app
Having unit tests is a must for any application.
The https://github.com/djangoflow/django-df-chat/issues/2 talks about a refactoring,
which could break the current flow of the chat app. So, with this test case we will be confident that the chat works the same as before(probably with some minute changes).

So, I have added the "df_chat.tests.TestChat" test case, to test the complete flow of our chat application - connection establishment, authentication and messaging.