Moving the "is_online" field of the RoomUser model onto a new UserChat model
The goal is to inform all rooms when a user is online, instead of just the current room they are in. To achieve this, we created a new model named 'UserChat' which holds attributes and fields that are common to all rooms. We then moved the 'is_online' field from the RoomUser model to the UserChat model.

To avoid disrupting users who are currently online, the migration process first creates the UserChat model and then creates UserChat objects for users who are currently online, setting their 'is_online' attribute to True. Finally, we delete the 'is_online' field from the RoomUser model.