Allowing users to connect to multiple rooms using a single websocket endpoint
Renamed RoomConsumer to RoomsConsumer, as it now allows users to connect to multiple rooms.
Once connected to the websocket, we automatically subscribe to all the rooms, the user is part of.
So, the user will now be able to listen to activities across his rooms, without having to use multiple connections.